Binned motif enrichment analysis.

Calculate motif enrichment analysis on bins of sequences. Motif occurrences in each bin are contrasted to the motif occurrences in a background set of sequences (by default the sequences in all other bins).

calcBinnedMotifEnr(
  seqs,
  bins,
  motifs,
  method = c("R", "Homer"),
  pseudocount.log2enr = 8,
  pseudocount.pearsonResid = 0.001,
  p.adjust.method = "BH",
  BPPARAM = SerialParam(),
  verbose = FALSE,
  ...
)

Arguments

seqs
The sequences to analyze. Either a DNAStringSet (for method = "R") or a GRanges object defining the genomic regions from which to extract the sequences (for method = "Homer").

bins
Factor of the same length and order as seqs, indicating the bin for each sequence or genomic region. Typically the return value of bin.

motifs
The motifs for which to calculate enrichments. Either a PWMatrixList (for method = "R") or a file with motifs in Homer format (for method = "Homer").

method
A character scalar specifying the backend to use for enrichment calculations. One of "R" (default) or "Homer".

pseudocount.log2enr
A numerical scalar with the pseudocount to add to foreground and background counts when calculating log2 motif enrichments

pseudocount.pearsonResid
A numerical scalar with the pseudocount to add to foreground and background frequencies when calculating expected counts and Pearson residuals.

p.adjust.method
A character scalar selecting the p value adjustment method (used in p.adjust).

BPPARAM
Specifies the number of CPU cores to use for parallel processing. Either a SerialParam or a MulticoreParam object instance, or for method = "Homer" a numeric scalar.

verbose
A logical scalar. If TRUE, print progress messages.

...
Additional arguments for passed to calcBinnedMotifEnrR or calcBinnedMotifEnrHomer.

Value

A SummarizedExperiment object with motifs in rows and bins in columns, containing six assays:

negLog10P: -log10 P values

negLog10Padj: -log10 adjusted P values

pearsonResid: motif enrichments as Pearson residuals

log2enr: motif enrichments as log2 ratios

sumForegroundWgtWithHits: Sum of foreground sequence weights in a bin that have motif hits

sumBackgroundWgtWithHits: Sum of background sequence weights in a bin that have motif hits

Details

This function is a wrapper for calcBinnedMotifEnrR or calcBinnedMotifEnrHomer. For additional supported arguments (via ...) please see the help pages of these functions.

See also