Cluster k-mers

Given enriched k-mers (typically the result of getKmerFreq), cluster k-mers into groups of similar k-mers that are likely to originate from the same motif.

clusterKmers(
  x,
  method = c("cooccurrence", "similarity"),
  allowReverseComplement = FALSE,
  nKmers = NULL,
  maxShift = NULL,
  minSim = NULL,
  seqs = NULL,
  zoops = TRUE,
  n = 1L
)

Arguments

x
Enriched k-mers, either a character vector or the return value of getKmerFreq, from which enriched k-mers can be extracted.

method
A character scalar specifying the clustering method to use. Currently, either "cooccurrence" (co-occurrence of k-mers in a set of sequences, the default) or "similarity" (using k-mer similarities, see Details below).

allowReverseComplement
A logical scalar. If TRUE, also include the reverse complement of enriched k-mers in the analysis. For method = "cooccurrence", this will extract also reverse complement k-mers from the co-occurrence count matrix for graph estimation. For method = "similarity", this will use the reverse complement of a k-mer if it has a higher toal similarity to all other k-mers than the original k-mer. In that case, the resulting k-mer is reported with the suffix _rc in the results, and the reverse-complement k-mer sequence is used to calculate pairwise similarities.

nKmers
A numeric scalar. If x is a list as returned by getKmerFreq, the number of top enriched k-mers to use ordered by padj. If NULL (the default), it will use nKmers <- max(10, sum(x$padj < 0.05)). nKmers is ignored if x is a character vector.

maxShift
(Only used for method = "similarity".) A numeric scalar with the maximal number of shifts to perform when calculating k-mer similarities. If NULL (the default), it will use maxShift <- kmerLen - 2.

minSim
(Only used for method = "similarity".) A numeric scalar with the minimal k-mer similarity to link two k-mer nodes in the graph by an edge. If NULL, it will use minSim <- kmerLen - maxShift + 1.

seqs
(Only used for method = "cooccurrence".) A DNAStringSet with the sequences, in which k-mer co-occurrences will be counted using countKmerPairs.

zoops
(Only used for method = "cooccurrence".) A logical scalar passed to countKmerPairs.

n
(Only used for method = "cooccurrence".) An integer scalar defining the maximum downstream distance of second k-mers, relative to the start position of the first k-mer (default: n = NULL).

Value

A named numeric vector, with k-mer as names and values indicating the k-mer cluster memberships. In attr(y, "graph") (where y is the return value), you can get the k-mer graph on which the clustering is based.

Details

The clustering is performed depending on method. For method = "cooccurrence", countKmerPairs(x = seqs, k = k, n = 1, zoops = zoops) will be used to first get a pairwise co-occurrence count matrix. Rows and columns corresponding to enriched k-mers from x (and their reverse complements, if allowReverseComplement = TRUE) will be extracted, and the resulting adjacency matrix will be converted to a graph, in which clusters will be identified as communities. For method = "similarity", all pairwise k-mer distances for all possible shifts (defined by maxShift) are calculated, defined as the Hamming distance of the overlapping substring plus the number of shifts. For each k-mer pair, the minimal distance over shifts is retained. If allowReverseComplement == TRUE, this procedure is repeated to compare each k-mer to all reverse-complemented k-mers, and replace it with the reverse-complemented version if this yields a lower sum of pairwise distances. The resulting distance matrix is then converted into a similarity matrix by subtracting it from the length of the k-mers. Elements less than minSim are set to zero, and the matrix is used as an adjacency matrix to construct a k-mer graph. The clusters are identified as the connected components in this graph.

See also

Examples

clusterKmers(x = c("AAA", "AAT", "GGG"),
             method = "similarity", minSim = 2)
#> clustering 3 3-mers (similarity-based, maxShift: 1, minSim: 2)
#> AAA AAT GGG 
#>   1   1   2 
#> attr(,"graph")
#> IGRAPH fcff215 UNW- 3 1 -- 
#> + attr: name (v/c), weight (e/n)
#> + edge from fcff215 (vertex names):
#> [1] AAA--AAT