Transform k-mer enrichments to motif enrichments.

Using a set of know motifs and the result of a k-mer enrichment analysis (k-mers by bins), transform the results into motif enrichments (motifs by bins).

convertKmersToMotifs(x, m, BPPARAM = SerialParam(), verbose = FALSE)

Arguments

x
A SummarizedExperiment with the results of a k-mer enrichment analysis (typically generated by calcBinnedKmerEnr).

m
Either a PFMatrixList, or a character scalar with a file containing motifs in HOMER format (loaded into a PFMatrixList by homerToPFMatrixList).

BPPARAM
An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation.

verbose
A logical scalar. If TRUE, report on progress.

See also