monaLisa - MOtif aNAlysis with Lisa.

monaLisa is a collection of tools that simplify motif enrichment analyses in genomic regions of interest.

Details

She makes use of her father Homer (http://homer.ucsd.edu/homer/index.html) and other algorithms to search for motif hits and look for enriched motifs in sets of genomic regions, compared to all other regions.

Known motifs can for example be obtained from a collection of transcription factor binding site specificities, such as JASPAR2020.

See also

Author

Dania Machlab, Lukas Burger, Charlotte Soneson and Michael Stadler