Randomized Lasso Stability Selection

This function runs randomized lasso stability selection as presented by Meinshausen and Bühlmann (2010) and with the improved error bounds introduced by Shah and Samworth (2013). The function uses the stabsel function from the stabs package, but implements the randomized lasso version.

randomized_stabsel(x = x, y = y, weakness = 0.8, cutoff = 0.8, PFER = 2, ...)

Arguments

x
the predictor matrix.

y
the response vector.

weakness
parameter that can take a value between 0 and 1. It affects how strict the method will be in selecting predictors. It is set to 0.8 by default. The closer it goes to 0 the more stringent the selection. A weakness value of 1 is identical to performing lasso stability selection (not the randomized version).

cutoff
value between 0 and 1 which is the cutoff for the selection probability. Any variable with a selection probability that is higher than the set cutoff will be selected. It is set to 0.8 by default.

PFER
the absolute number of false positives that we allow for in the final list of selected variables. For details see Meinshausen and Bühlmann (2010).

...
additional parameters for glmnet.

Value

a stabsel object. It contains the following:

phat - a matrix containing selection probabilities, where columns are regularization steps and rows are predictors.

selected - a vector containing the selected predictors.

max - maximum of selection probabilities.

cutoff - the selection probability cutoff.

q - average number of selected variables used.

PFER - the realized upper bound for the per-family error rate (number of falsely selected predictors among the group of selected predictors).

specifiedPFER - the set upper bound for the per-family error rate (number of falsely selected predictors among the group of selected predictors).

p - the total number of predictors.

B - the number of subsamples.

sampling.type - sampling type used for stability selection.

assumption - assumptions made on the selection probabilities.

call - the call.

Details

Randomized lasso stability selection runs randomized lasso several times on subsamples of the response variable and predictor matrix (stability selection). N/2 elements from the response variable are randomly chosen, where N is the length of the vector. Their corresponsing section of the predictor matrix is also chosen, and the glmnet.randomized_lasso function is applied. This is done multiple times, and results in selection probabilities for each predictor. The probability of a specific predictor is the number of times it was selected divided by the total number of subsamples that were done (total number of times the regression was performed).

We make use of the stabs package that implements lasso stability selection, and adapt it to run the randomized lasso stability selection. The output is an object of type stabsel.

References

N. Meinshausen and P. Bühlmann (2010), Stability Selection, Journal of the Royal Statistical Society: Series B (Statistical Methodology), 72, 417–73.
R.D. Shah and R.J. Samworth (2013), Variable Selection with Error Control: Another Look at Stability Selection, Journal of the Royal Statistical Society: Series B (Statistical Methodology), 75, 55–80.
B. Hofner, L. Boccuto, and M. Göker (2015), Controlling False Discoveries in High-Dimensional Situations: Boosting with Stability Selection, BMC Bioinformatics, 16 144.

See also