Prepare and run HOMER motif enrichment analysis.

Run complete HOMER motif enrichment analysis, consisting of calls to prepareHomer, system and parseHomerOutput.

runHomer(
  gr,
  b,
  genomedir,
  outdir,
  motifFile,
  homerfile = findHomer(),
  regionsize = "given",
  Ncpu = 2L
)

Arguments

gr
A GRanges object (or an object that can be coerced to one) with the genomic regions to analyze.

b
A vector of the same length as gr that groups its elements into bins (typically a factor, such as the one returned by bin).

genomedir
Directory containing sequence files in Fasta format (one per chromosome).

outdir
A path specifying the folder into which the output files will be written.

motifFile
A file with HOMER formatted PWMs to be used in the enrichment analysis.

homerfile
Path and file name of the findMotifsGenome.pl HOMER script.

regionsize
The peak size to use in HOMER ("given" keeps the coordinate region, an integer value will keep only that many bases in the region center).

Ncpu
Number of parallel threads that HOMER can use.

Value

A SummarizedExperiment x with

assays(x)

containing the four components p, FDR, enr and log2enr), each a motif (rows) by bin (columns) matrix with raw -log10 P values, -log10 false discovery rates and motif enrichments as Pearson residuals (enr) and as log2 ratios (log2enr).

rowData(x)

containing information about the motifs.

colData(x)

containing information about the bins.

metaData(x)

containing meta data on the object (e.g. parameter values and distances between pairs of motifs).

See also