Prepare and run HOMER motif enrichment analysis.

Run complete HOMER motif enrichment analysis, consisting of calls to prepareHomer, system2 and parseHomerOutput.

calcBinnedMotifEnrHomer(
  gr,
  b,
  genomedir,
  outdir,
  motifFile,
  homerfile = findHomer(),
  regionsize = "given",
  pseudocount.log2enr = 8,
  pseudocount.pearsonResid = 0.001,
  p.adjust.method = "BH",
  Ncpu = 2L,
  verbose = FALSE
)

Arguments

gr
A GRanges object (or an object that can be coerced to one) with the genomic regions to analyze.

b
A vector of the same length as gr that groups its elements into bins (typically a factor, such as the one returned by bin).

genomedir
Directory containing sequence files in Fasta format (one per chromosome).

outdir
A path specifying the folder into which the output files will be written.

motifFile
A file with HOMER formatted PWMs to be used in the enrichment analysis.

homerfile
Path and file name of the findMotifsGenome.pl HOMER script.

regionsize
The peak size to use in HOMER ("given" keeps the coordinate region, an integer value will keep only that many bases in the region center).

pseudocount.log2enr
A numerical scalar with the pseudocount to add to foreground and background counts when calculating log2 motif enrichments

pseudocount.pearsonResid
A numerical scalar with the pseudocount to add to foreground and background frequencies when calculating expected counts and Pearson residuals.

p.adjust.method
A character scalar selecting the p value adjustment method (used in p.adjust).

Ncpu
Number of parallel threads that HOMER can use.

verbose
A logical scalar. If TRUE, print progress messages.

Value

A SummarizedExperiment object with motifs in rows and bins in columns, containing six assays:

negLog10P: -log10 P values

negLog10Padj: -log10 adjusted P values

pearsonResid: motif enrichments as Pearson residuals

log2enr: motif enrichments as log2 ratios

sumForegroundWgtWithHits: Sum of foreground sequence weights in a bin that have motif hits

sumBackgroundWgtWithHits: Sum of background sequence weights in a bin that have motif hits

See also