Calculate observed and expected k-mer frequencies

Given a set of sequences, calculate observed and expected k-mer frequencies. Expected frequencies are based on a Markov model of order MMorder.

getKmerFreq(
  seqs,
  kmerLen = 5,
  MMorder = 1,
  pseudoCount = 1,
  zoops = TRUE,
  strata = rep(1L, length(seqs))
)

Arguments

seqs
Set of sequences, either a character vector or a DNAStringSet.

kmerLen
A numeric scalar giving the k-mer length.

MMorder
A numeric scalar giving the order of the Markov model used to calculate the expected frequencies.

pseudoCount
A numeric scalar - will be added to the observed counts for each k-mer to avoid zero values.

zoops
A logical scalar. If TRUE (the default), only one or zero occurences of a k-mer are considered per sequence.

strata
A factor or a numeric scalar defining the strata of sequences. A separate Markov model and expected k-mer frequencies are estimated for the set of sequences in each stratum (level in a strata factor). If strata is a scalar value, it will be interpreted as the number of strata to split the sequences into according to their CpG observed-over-expected counts using kmeans(CpGoe, centers = strata).

Value

A list with observed and expected k-mer frequencies (freq.obs and freq.exp, respectively), and enrichment statistics for each k-mer.