Calculate similarities between motifs and k-mers.

For each motif, calculate it's similarity to all k-mers of length kmerLen, defined as the maximal probability of observing the k-mer given the base frequencies of the motif (the maximum is taken over for all possible ungapped alignments between motif and k-mer). If necessary matrices are padded on the sides with background base frequencies (assuming all bases to have a frequency of 0.25).

motifKmerSimilarity(x, kmerLen = 5, BPPARAM = SerialParam(), verbose = FALSE)

Arguments

x
Either a PFMatrixList, or a character scalar with a file containing motifs in HOMER format (used directly method = "HOMER", loaded into a PFMatrixList by homerToPFMatrixList for method = "R").

kmerLen
A numeric scalar giving the k-mer length.

BPPARAM
An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation.

verbose
A logical scalar. If TRUE, report on progress.

Value

A matrix of probabilties for each motif - k-mer pair.

See also